Refactor note
